+++ src/app/pages/article/article.component.ts
@@ -7,7 +7,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'page-acticle',
+  selector: 'page-article',
   template: `<router-outlet></router-outlet>`
 })
 export class ArticleComponent {