+++ src/app/pages/example/components/buttons/components/dropdownButtons/dropdownButtons.component.ts
@@ -1,4 +1,4 @@
-import { Component, HostListener } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'box-dropdown-buttons',