+++ src/app/pages/pages.utils.ts
@@ -21,16 +21,16 @@ interface IItemSelectChangeResult {
 }
 
 // 合并 form 到实例本身
-export function mergeFormControlsToInstance(instance, form) {
+export const mergeFormControlsToInstance = (instance, form) => {
   if (form instanceof FormGroup) {
     Object.keys(form.controls).forEach(keyword => {
       instance[keyword] = form.controls[keyword];
     });
   }
-}
+};
 
 // 对批量操作进行更新操作
-export function handleBatchSelectChange(options: ISelectChangeOptions): TSelectedIds {
+export const handleBatchSelectChange = (options: ISelectChangeOptions): TSelectedIds => {
   const { data, isSelect } = options;
   if (!data.length) {
     return;
@@ -38,20 +38,20 @@ export function handleBatchSelectChange(options: ISelectChangeOptions): TSelecte
   data.forEach(item => (item.selected = isSelect));
   options.selectedIds = isSelect ? data.map(item => item._id) : [];
   return options.selectedIds;
-}
+};
 
 // 对单个勾选进行更新操作
-export function handleItemSelectChange(options: ISelectChangeOptions): IItemSelectChangeResult {
+export const handleItemSelectChange = (options: ISelectChangeOptions): IItemSelectChangeResult => {
   const { data } = options;
   options.selectedIds = data.filter(item => item.selected).map(item => item._id);
   return {
     selectedIds: options.selectedIds,
     all: options.selectedIds.length === data.length
   };
-}
+};
 
 // 表单验证
-export function formControlStateClass(control: AbstractControl, errClassName?: string, isSubmited?: boolean): string {
+export const formControlStateClass = (control: AbstractControl, errClassName?: string, isSubmited?: boolean): string => {
   if (control.touched || control.root.touched || control.dirty || control.root.dirty || isSubmited) {
     if (control.valid) {
       return 'has-success';
@@ -59,4 +59,4 @@ export function formControlStateClass(control: AbstractControl, errClassName?: s
       return errClassName || 'has-error';
     }
   }
-}
+};