+++ src/app/transformers/gravatar.ts
@@ -6,8 +6,7 @@
 import * as gravatar from 'gravatar';
 import { GRAVATAR_API } from '@/config';
 
-export function getGravatar(email: string): string {
+export const getGravatar = (email: string): string => {
   const gravatar_url = gravatar.url(email, { protocol: 'https' });
   return gravatar_url.replace('https://s.gravatar.com/avatar', GRAVATAR_API);
-}
-
+};